+++ Ekom/Models/IProduct.cs
@@ -45,7 +45,7 @@ namespace Ekom.Models
         /// <value>
         /// The stock.
         /// </value>
-        int Stock { get; }
+        decimal Stock { get; }
 
         /// <summary>
         /// Gets the availability of the product and variants.
